backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
apscheduler
